silence compiler warning when converting MAXBIGINT to a double

+++ src/variable.cpp
@@ -2788,8 +2788,8 @@ double Variable::collapse_tree(Tree *tree)
     else if (update->ntimestep < ivalue2) {
       int offset = update->ntimestep - ivalue1;
       tree->value = ivalue1 + (offset/ivalue3)*ivalue3 + ivalue3;
-      if (tree->value > ivalue2) tree->value = MAXBIGINT;
-    } else tree->value = MAXBIGINT;
+      if (tree->value > ivalue2) tree->value = (double) MAXBIGINT;
+    } else tree->value = (double) MAXBIGINT;
     return tree->value;
   }
 
@@ -3122,8 +3122,8 @@ double Variable::eval_tree(Tree *tree, int i)
     else if (update->ntimestep < ivalue2) {
       int offset = update->ntimestep - ivalue1;
       arg = ivalue1 + (offset/ivalue3)*ivalue3 + ivalue3;
-      if (arg > ivalue2) arg = MAXBIGINT;
-    } else arg = MAXBIGINT;
+      if (arg > ivalue2) arg = (double) MAXBIGINT;
+    } else arg = (double) MAXBIGINT;
     return arg;
   }
 
@@ -3699,8 +3699,8 @@ int Variable::math_function(char *word, char *contents, Tree **tree,
       else if (update->ntimestep < ivalue2) {
         int offset = update->ntimestep - ivalue1;
         value = ivalue1 + (offset/ivalue3)*ivalue3 + ivalue3;
-        if (value > ivalue2) value = MAXBIGINT;
-      } else value = MAXBIGINT;
+        if (value > ivalue2) value = (double) MAXBIGINT;
+      } else value = (double) MAXBIGINT;
       argstack[nargstack++] = value;
     }
